AttendanceManagement/requirements.txt
flask
numpy
opencv-contrib-python
opencv-python
openpyxl
pandas
pillow
pyttsx3
